requirements.txt
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==5.0.2
Django==4.2.8
django-cleanup==8.1.0
django-decouple==2.1
django-jazzmin==2.6.0
django-router==1.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
